+++ src/g_heretic/a_hereticweaps.cpp
@@ -462,8 +462,8 @@ DEFINE_ACTION_FUNCTION(AActor, A_MacePL1Check)
 	self->velx = FixedMul(7*FRACUNIT, finecosine[angle]);
 	self->vely = FixedMul(7*FRACUNIT, finesine[angle]);
 #else
-	double velscale = sqrtf ((float)self->velx * (float)self->velx +
-							 (float)self->vely * (float)self->vely);
+	double velscale = sqrt ((double)self->velx * (double)self->velx +
+							 (double)self->vely * (double)self->vely);
 	velscale = 458752 / velscale;
 	self->velx = (int)(self->velx * velscale);
 	self->vely = (int)(self->vely * velscale);
